+++ test/src/test_storage_xyz.cpp
@@ -9,9 +9,7 @@
 
 #include <gtest/gtest.h>
 
-// #include <kodo/shallow_storage_decoder.hpp>
-#include <kodo/deep_storage_decoder.hpp>
-// #include <kodo/storage_encoder.hpp>
+
 #include <kodo/rfc5052_partitioning_scheme.hpp>
 #include <kodo/rlnc/full_rlnc_codes.hpp>
 #include <kodo/partial_shallow_symbol_storage.hpp>